+++ react-native/scripts/initializeSubmodule.sh
@@ -10,8 +10,8 @@ GRAY='\033[1;30m'
 NC='\033[0m' # No Color
 
 echo -e "${BLUE}Initializing Showcase Submodule...${NC}"
-if [ ! -f ./demos/showcase/package.json ];
-    then git submodule init && git submodule update;
+if [ ! -f ./demos/showcase/package.json ]; 
+    then git submodule init && git submodule update; 
     else echo -e "${BBLUE}Already initialized${NC}" && exit 0;
 fi
 echo -e "${BBLUE}Initialization Complete${NC}\r\n"